Segregate Dashboard into components
